QueryInputInteractor connection tests
