+++ Tests/Sources/QueryInput/TestQueryInputController.swift
@@ -1,0 +1,33 @@
+//
+//  TestQueryInputController.swift
+//  InstantSearchCore
+//
+//  Created by Vladislav Fitc on 04/12/2019.
+//  Copyright © 2019 Algolia. All rights reserved.
+//
+
+import Foundation
+import InstantSearchCore
+
+class TestQueryInputController: QueryInputController {
+  
+  var query: String? {
+    didSet {
+      guard oldValue != query else { return }
+      onQueryChanged?(query)
+    }
+  }
+  
+  var onQueryChanged: ((String?) -> Void)?
+  var onQuerySubmitted: ((String?) -> Void)?
+  
+  func setQuery(_ query: String?) {
+    self.query = query
+  }
+  
+  func submitQuery() {
+    onQuerySubmitted?(query)
+  }
+  
+  
+}
